requirements.txt
faiss_cpu==1.8.0
fitz==0.0.1.dev2
Flask==3.0.3
groq==0.9.0
numpy==2.0.1
Pillow==10.0.1
Pillow==10.4.0
sentence_transformers==2.7.0
Werkzeug==3.0.3
